move scalar into separate module
